refactor: 이벤트 리스너 위치 변경
