+++ src/main/java/com/example/oatnote/event/UserRegisteredEvent.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.event.model;
+package com.example.oatnote.event;
 
 public record UserRegisteredEvent(
     String userId
